+++ src/trx/game/level/format/format_tr1.c
@@ -145,6 +145,7 @@ static bool M_Load(const LEVEL_FORMAT_LOADER *const loader, VFILE *const file)
     }
 
     Level_Section_ReadCamerasAndSinks(ctx, file);
+    Level_Section_ReadFlybyCameras(ctx, file);
     Level_Section_ReadSoundSources(ctx, file);
     Level_Section_ReadPathingData(ctx, file);
     Level_Section_ReadAnimatedTextureRanges(ctx, file);