+++ src/trx/game/level/sections/cinematics.c
@@ -63,14 +63,29 @@ void Level_Section_ReadCamerasAndSinks(
 
 void Level_Section_ReadFlybyCameras(LEVEL_CONTEXT *const ctx, VFILE *const file)
 {
+    const int32_t inject_count = Inject_GetDataCount(IDT_FLYBY_CAMERAS);
+    if (ctx->loader->game_version < 4) {
+        Camera_InitialiseFlybys(inject_count);
+        return;
+    }
+
     BENCHMARK benchmark = Benchmark_Start();
     const int16_t num_cameras = VFile_ReadS16(file);
+    LEVEL_CONTEXT_INFO *const info = &ctx->info;
+    info->cameras.flyby_count = num_cameras;
     VFile_Skip(file, sizeof(int16_t)); // reserved padding
-    Camera_InitialiseFlybys(num_cameras); // TODO: detect injections
+    Camera_InitialiseFlybys(num_cameras + inject_count);
     LOG_INFO("flyby cameras: %d", num_cameras);
+    Level_Section_AppendFlybyCameras(0, num_cameras, file);
 
+    Benchmark_End(&benchmark, nullptr);
+}
+
+void Level_Section_AppendFlybyCameras(
+    const int32_t base_idx, const int32_t num_cameras, VFILE *const file)
+{
     for (int32_t i = 0; i < num_cameras; i++) {
-        FLYBY_CAMERA *const camera = Camera_GetFlybyCamera(i);
+        FLYBY_CAMERA *const camera = Camera_GetFlybyCamera(base_idx + i);
         M_ReadPosition(&camera->pos, file);
         M_ReadPosition(&camera->target, file);
         camera->sequence = VFile_ReadU8(file);
@@ -102,8 +117,6 @@ void Level_Section_ReadFlybyCameras(LEVEL_CONTEXT *const ctx, VFILE *const file)
         camera->flags.one_shot         = (flags & 0x8000) != 0;
         // clang-format on
     }
-
-    Benchmark_End(&benchmark, nullptr);
 }
 
 void Level_Section_ReadDemoData(LEVEL_CONTEXT *const ctx, VFILE *const file)